add PartnerForm
