+++ src/components/admin-page/Partners/PartnerList/PartnerList.jsx
@@ -4,10 +4,22 @@ import PartnerCard from '@/src/components/shared/PartnerCard/PartnerCard'
 import MainButton from '@/src/components/shared/MainButton/MainButton'
 import { Icon } from '@/src/components/shared/Icon/Icon'
 import styles from './PartnerList.module.scss'
+import { useState } from 'react'
+import AdminModal from '@/src/components/modals/AdminModal/AdminModal'
 
-export default function PartnerList({data}) {
+export default function PartnerList({data, hendleRemove}) {
   const router = useRouter();
+  const[ idPartner, setIdPartner ] = useState(null)
 
+  const closeModal=()=>{
+    setIdPartner(null)
+  }
+
+  const okRemove=()=>{
+    hendleRemove(idPartner)
+    setIdPartner(null)
+  }
+  
   return ( <>
     {data?.length ? <ul className={styles.list}>
       {data.map((el)=>{
@@ -20,7 +32,7 @@ export default function PartnerList({data}) {
               <Icon  width={24} height={24} name='edit'/>
             </MainButton>
 
-            <MainButton variant='admin' className={styles.btn}>
+            <MainButton variant='admin' onClick={()=>{setIdPartner(el._id)}} className={styles.btn}>
               <Icon width={24} height={24} name='remove'/>
             </MainButton>
           </div>
@@ -28,6 +40,8 @@ export default function PartnerList({data}) {
       })}
     </ul>: null
     }
+
+    <AdminModal isOpen={idPartner} handleCallback={closeModal} handleOkCallback={okRemove} title={'Ви впевнені, що хочете видалити партнера?'} btnBlok={true}></AdminModal>
   </>
   )
 }
