+++ src/components/admin-page/Partners/Partners.jsx
@@ -1,20 +1,34 @@
 'use client';
 import styles from './Partners.module.scss'
 import { useRouter } from '@/src/navigation';
-import { useQuery } from '@tanstack/react-query';
-import { getAllPartners } from '@/src/api/partners';
+import { useMutation, useQuery } from '@tanstack/react-query';
+import { deletePartnerById, getAllPartners } from '@/src/api/partners';
 import SectionAdmin from "../SectionAdmin/SectionAdmin";
 import MainButton from '../../shared/MainButton/MainButton';
 import PartnerList from './PartnerList/PartnerList';
 import { useState } from 'react';
 import { Icon } from '../../shared/Icon/Icon';
-//PartnerCard edit-partner
+import Loader from '../../shared/loader/Loader';
+import UseAlert from '../../shared/UseAlert/UseAlert';
+import stateUseAlert from '@/src/state/stateUseAlert';
+
 export default function Partners() {
   const router = useRouter();
   const [ search, setSearch ] = useState('')
+  const open = stateUseAlert(state => state.open);
 
   const { isError, data, refetch } = useQuery({ queryKey: ['partners', search], 
     queryFn:()=>{return getAllPartners({query:search})}, keepPreviousData: true });
+
+  const deletePartner = useMutation({
+    mutationFn:(id) => {
+      return deletePartnerById(id)
+    },onSuccess: () => {
+      refetch()
+    },onError:()=>{
+      open('error', false)
+    }})
+
   return(
     <SectionAdmin title={'Партнери'} hendleSearch={setSearch} lang={true}>
       <MainButton  variant='admin' className={styles.btn} onClick={()=>{
@@ -23,7 +37,11 @@ export default function Partners() {
         <Icon name={'plus_icon'} width={24} height={24}/>
         {'Додати партнера'}</MainButton >
 
-      {data?.results && !isError && <PartnerList data={data?.results}/>}
+      {data?.results && !isError && <PartnerList data={data?.results} hendleRemove={deletePartner.mutate}/>
+      }
+
+      {deletePartner.isPending && <Loader/>}
+      <UseAlert/>
     </SectionAdmin>
   )
 }
